refactor: changed datatype from int to double

+++ common/src/main/java/et/edu/astu/common/interfaces/TransactionResponse.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 
 public interface TransactionResponse {
     String getTransactionId();
-    Integer getAmount();
+    Double getAmount();
     String getType();
     LocalDateTime getCreatedAt();
 }
